+++ pybamm/expression_tree/binary_operators.py
@@ -162,12 +162,6 @@ class BinaryOperator(pybamm.Symbol):
         """ Calculate the jacobian of a binary operator. """
         raise NotImplementedError
 
-    def _binary_simplify(self, new_left, new_right):
-        """ Simplify a binary operator """
-        return pybamm.simplify_if_constant(
-            self._binary_new_copy(new_left, new_right), clear_domains=False
-        )
-
     def _binary_evaluate(self, left, right):
         """ Perform binary operation on nodes 'left' and 'right'. """
         raise NotImplementedError
@@ -226,10 +220,6 @@ class Power(BinaryOperator):
         with np.errstate(invalid="ignore"):
             return left ** right
 
-    def _binary_simplify(self, new_left, new_right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplified_power(new_left, new_right)
-
 
 class Addition(BinaryOperator):
     """A node in the expression tree representing an addition operator
@@ -252,10 +242,6 @@ class Addition(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left + right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_addition_subtraction(self.__class__, left, right)
 
 
 class Subtraction(BinaryOperator):
@@ -280,12 +266,6 @@ class Subtraction(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left - right
-
-    def _binary_simplify(self, left, right):
-        """
-        See :meth:`pybamm.BinaryOperator._binary_simplify()`.
-        """
-        return pybamm.simplify_addition_subtraction(self.__class__, left, right)
 
 
 class Multiplication(BinaryOperator):
@@ -332,10 +312,6 @@ class Multiplication(BinaryOperator):
         else:
             return left * right
 
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
-
 
 class MatrixMultiplication(BinaryOperator):
     """A node in the expression tree representing a matrix multiplication operator
@@ -377,10 +353,6 @@ class MatrixMultiplication(BinaryOperator):
     def _binary_evaluate(self, left, right):
         """ See :meth:`pybamm.BinaryOperator._binary_evaluate()`. """
         return left @ right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
 
 
 class Division(BinaryOperator):
@@ -424,10 +396,6 @@ class Division(BinaryOperator):
                     return left * np.inf
             else:
                 return left / right
-
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
 
 
 class Inner(BinaryOperator):
@@ -486,10 +454,6 @@ class Inner(BinaryOperator):
         """ See :meth:`pybamm.BinaryOperator._binary_new_copy()`. """
         return pybamm.inner(left, right)
 
-    def _binary_simplify(self, left, right):
-        """ See :meth:`pybamm.BinaryOperator._binary_simplify()`. """
-        return pybamm.simplify_multiplication_division(self.__class__, left, right)
-
     def _evaluates_on_edges(self, dimension):
         """ See :meth:`pybamm.Symbol._evaluates_on_edges()`. """
         return False
@@ -516,7 +480,7 @@ def inner(left, right):
     if pybamm.is_scalar_one(right):
         return left
 
-    return pybamm.simplify_if_constant(pybamm.Inner(left, right), clear_domains=False)
+    return pybamm.simplify_if_constant(pybamm.Inner(left, right))
 
 
 class Heaviside(BinaryOperator):
@@ -748,7 +712,7 @@ def simplified_power(left, right):
             if new_left.is_constant() or new_right.is_constant():
                 return new_left / new_right
 
-    return pybamm.simplify_if_constant(pybamm.Power(left, right), clear_domains=False)
+    return pybamm.simplify_if_constant(pybamm.Power(left, right))
 
 
 def simplified_addition(left, right):
@@ -820,9 +784,7 @@ def simplified_addition(left, right):
             new_sum.copy_domains(pybamm.Addition(left, right))
             return new_sum
 
-    return pybamm.simplify_if_constant(
-        pybamm.Addition(left, right), clear_domains=False
-    )
+    return pybamm.simplify_if_constant(pybamm.Addition(left, right))
 
 
 def simplified_subtraction(left, right):
@@ -880,9 +842,7 @@ def simplified_subtraction(left, right):
     if left.id == right.id:
         return pybamm.zeros_like(left)
 
-    return pybamm.simplify_if_constant(
-        pybamm.Subtraction(left, right), clear_domains=False
-    )
+    return pybamm.simplify_if_constant(pybamm.Subtraction(left, right))
 
 
 def simplified_multiplication(left, right):
@@ -929,9 +889,7 @@ def simplified_multiplication(left, right):
 
     # Return constant if both sides are constant
     if left.is_constant() and right.is_constant():
-        return pybamm.simplify_if_constant(
-            pybamm.Multiplication(left, right), clear_domains=False
-        )
+        return pybamm.simplify_if_constant(pybamm.Multiplication(left, right))
 
     # Simplify (B @ c) * a to (a * B) @ c if (a * B) is constant
     # This is a common construction that appears from discretisation of spatial
@@ -1060,9 +1018,7 @@ def simplified_division(left, right):
             if new_right.is_constant():
                 return l_left * new_right
 
-    return pybamm.simplify_if_constant(
-        pybamm.Division(left, right), clear_domains=False
-    )
+    return pybamm.simplify_if_constant(pybamm.Division(left, right))
 
 
 def simplified_matrix_multiplication(left, right):
@@ -1116,9 +1072,7 @@ def simplified_matrix_multiplication(left, right):
             r_left, r_right = right.orphans
             return (left @ r_left) + (left @ r_right)
 
-    return pybamm.simplify_if_constant(
-        pybamm.MatrixMultiplication(left, right), clear_domains=False
-    )
+    return pybamm.simplify_if_constant(pybamm.MatrixMultiplication(left, right))
 
 
 def minimum(left, right):
@@ -1133,7 +1087,7 @@ def minimum(left, right):
         out = Minimum(left, right)
     else:
         out = pybamm.softminus(left, right, k)
-    return pybamm.simplify_if_constant(out, clear_domains=False)
+    return pybamm.simplify_if_constant(out)
 
 
 def maximum(left, right):
@@ -1148,7 +1102,7 @@ def maximum(left, right):
         out = Maximum(left, right)
     else:
         out = pybamm.softplus(left, right, k)
-    return pybamm.simplify_if_constant(out, clear_domains=False)
+    return pybamm.simplify_if_constant(out)
 
 
 def softminus(left, right, k):