+++ pybamm/models/full_battery_models/lithium_ion/basic_dfn_half_cell.py
@@ -428,7 +428,6 @@ class BasicDFNHalfCell(BaseModel):
     def new_copy(self, build=False):
         new_model = self.__class__(name=self.name, options=self.options)
         new_model.use_jacobian = self.use_jacobian
-        new_model.use_simplify = self.use_simplify
         new_model.convert_to_format = self.convert_to_format
         new_model.timescale = self.timescale
         new_model.length_scales = self.length_scales